+++ rocketpy/mathutils/function.py
@@ -25,6 +25,7 @@ from scipy.interpolate import (
     RBFInterpolator,
     RegularGridInterpolator,
 )
+from scipy.spatial import Delaunay
 
 from rocketpy.plots.plot_helpers import show_or_save_plot
 from rocketpy.tools import deprecated, from_hex_decode, to_hex_encode
@@ -510,7 +511,9 @@ class Function:  # pylint: disable=too-many-public-methods
             case 0:  # linear
                 if self.__dom_dim__ == 1:
 
-                    def linear_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                    def linear_interpolation(
+                        x, x_min, x_max, x_data, y_data, coeffs
+                    ):  # pylint: disable=unused-argument
                         x_interval = bisect_left(x_data, x)
                         x_left = x_data[x_interval - 1]
                         y_left = y_data[x_interval - 1]
@@ -519,23 +522,31 @@ class Function:  # pylint: disable=too-many-public-methods
                         return (x - x_left) * (dy / dx) + y_left
 
                 else:
-                    interpolator = LinearNDInterpolator(self._domain, self._image)
+                    tri = Delaunay(self._domain)
+                    interpolator = LinearNDInterpolator(tri, self._image)
+                    self._nd_triangulation = tri
 
-                    def linear_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                    def linear_interpolation(
+                        x, x_min, x_max, x_data, y_data, coeffs
+                    ):  # pylint: disable=unused-argument
                         return interpolator(x)
 
                 self._interpolation_func = linear_interpolation
 
             case 1:  # polynomial
 
-                def polynomial_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                def polynomial_interpolation(
+                    x, x_min, x_max, x_data, y_data, coeffs
+                ):  # pylint: disable=unused-argument
                     return np.sum(coeffs * x ** np.arange(len(coeffs)))
 
                 self._interpolation_func = polynomial_interpolation
 
             case 2:  # akima
 
-                def akima_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                def akima_interpolation(
+                    x, x_min, x_max, x_data, y_data, coeffs
+                ):  # pylint: disable=unused-argument
                     x_interval = bisect_left(x_data, x)
                     x_interval = x_interval if x_interval != 0 else 1
                     a = coeffs[4 * x_interval - 4 : 4 * x_interval]
@@ -545,7 +556,9 @@ class Function:  # pylint: disable=too-many-public-methods
 
             case 3:  # spline
 
-                def spline_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                def spline_interpolation(
+                    x, x_min, x_max, x_data, y_data, coeffs
+                ):  # pylint: disable=unused-argument
                     x_interval = bisect_left(x_data, x)
                     x_interval = max(x_interval, 1)
                     a = coeffs[:, x_interval - 1]
@@ -581,7 +594,9 @@ class Function:  # pylint: disable=too-many-public-methods
             case 5:  # RBF
                 interpolator = RBFInterpolator(self._domain, self._image, neighbors=100)
 
-                def rbf_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                def rbf_interpolation(
+                    x, x_min, x_max, x_data, y_data, coeffs
+                ):  # pylint: disable=unused-argument
                     return interpolator(x)
 
                 self._interpolation_func = rbf_interpolation
@@ -602,7 +617,9 @@ class Function:  # pylint: disable=too-many-public-methods
                 # Store so extrapolation funcs can reuse it
                 self._grid_interpolator = grid_interpolator
 
-                def grid_interpolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                def grid_interpolation(
+                    x, x_min, x_max, x_data, y_data, coeffs
+                ):  # pylint: disable=unused-argument
                     return grid_interpolator(x)
 
                 self._interpolation_func = grid_interpolation
@@ -622,7 +639,9 @@ class Function:  # pylint: disable=too-many-public-methods
         match extrapolation:
             case 0:  # zero
 
-                def zero_extrapolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                def zero_extrapolation(
+                    x, x_min, x_max, x_data, y_data, coeffs
+                ):  # pylint: disable=unused-argument
                     return 0
 
                 self._extrapolation_func = zero_extrapolation
@@ -739,14 +758,18 @@ class Function:  # pylint: disable=too-many-public-methods
             case 2:  # constant
                 if self.__dom_dim__ == 1:
 
-                    def constant_extrapolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                    def constant_extrapolation(
+                        x, x_min, x_max, x_data, y_data, coeffs
+                    ):  # pylint: disable=unused-argument
                         return y_data[0] if x < x_min else y_data[-1]
 
                 elif self.__interpolation__ == "regular_grid":
                     grid_axes = self._grid_axes
                     grid_interpolator_const = self._grid_interpolator
 
-                    def constant_extrapolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                    def constant_extrapolation(
+                        x, x_min, x_max, x_data, y_data, coeffs
+                    ):  # pylint: disable=unused-argument
                         # Clamp each coordinate to its axis bounds, then interpolate
                         x_clamped = np.copy(x)
                         for i, axis in enumerate(grid_axes):
@@ -758,7 +781,9 @@ class Function:  # pylint: disable=too-many-public-methods
                 else:
                     extrapolator = NearestNDInterpolator(self._domain, self._image)
 
-                    def constant_extrapolation(x, x_min, x_max, x_data, y_data, coeffs):  # pylint: disable=unused-argument
+                    def constant_extrapolation(
+                        x, x_min, x_max, x_data, y_data, coeffs
+                    ):  # pylint: disable=unused-argument
                         return extrapolator(x)
 
                 self._extrapolation_func = constant_extrapolation
@@ -827,8 +852,11 @@ class Function:  # pylint: disable=too-many-public-methods
         min_domain = self._domain.T.min(axis=1)
         max_domain = self._domain.T.max(axis=1)
 
-        lower, upper = args < min_domain, args > max_domain
-        extrap = np.logical_or(lower.any(axis=1), upper.any(axis=1))
+        if hasattr(self, "_nd_triangulation"):
+            extrap = self._nd_triangulation.find_simplex(args) < 0
+        else:
+            lower, upper = args < min_domain, args > max_domain
+            extrap = np.logical_or(lower.any(axis=1), upper.any(axis=1))
 
         if extrap.any():
             result[extrap] = self._extrapolation_func(